Singleton-ized, List of changes:
- Moved everything to poisins.core namespace (and under folder core)
- renamed FileBrowserScript to FileBrowserView
- added a new static class FileBrowser (namespace poinins) and moved
  OpenFileBrowser() method there. This class also handles the callback
- changed callback functionality to a delagate
- optimized the prefabs (unmarked Raycast Target and Rich Text where
  possible, it was dropping performance quite a bit)
- updated example scenes and the Example.cs accordingly

overall the usage of this plugin came down to 2 steps:
1. keep the FileBrowserWindow.prefab somewhere in the scene
2. call poisins.FileBrowser.OpenFileBrowser( BrowserCallback );

NOTE: BrowserCallback is a delegate method with string argument
