+++ Assets/UnityUIFileBrowser/Scripts/Core.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 6169c654f39de0649ae71509f725006a
+folderAsset: yes
+timeCreated: 1482929085
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
