+++ Assets/UnityUIFileBrowser/Scripts/Core/DiskContentItemScript.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 1ea192e7c3bd471439a50e27d6d51572
-timeCreated: 1458775522
+guid: 83edc24603842194e93b220cd026d11b
+timeCreated: 1482929085
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
