+++ Assets/UnityUIFileBrowser/Scripts/Core/FileBrowserView.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 2123422bba68ab54ea66d6e5549eda6f
+timeCreated: 1482929085
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
