+++ Assets/UnityUIFileBrowser/Scripts/Core/ResetScale.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: b270ddeb22fc7d04ba2b448ee667e4da
+timeCreated: 1482929085
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
