--- Assets/UnityUIFileBrowser/Scripts/DiskContentItemScript.cs.meta
@@ -1,16 +1,0 @@
-fileFormatVersion: 2
-guid: 856b055b3b134204d956321e2b9d4817
-timeCreated: 1458773966
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences:
-  - IconField: {instanceID: 0}
-  - NameField: {instanceID: 0}
-  - FolderIcon: {fileID: 21300000, guid: a4e871a05709cc8449a349f743d8e8b5, type: 3}
-  - FileIcon: {fileID: 21300000, guid: 0e51a9163e43b2340ac7d28679f6b149, type: 3}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
