+++ Assets/UnityUIFileBrowser/Scripts/FileBrowser.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: fef473deaf668c84f848eefb415d2c04
-timeCreated: 1458772290
+guid: dbe2b05fed1e236498453c1e86ee8f7d
+timeCreated: 1482928184
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
