--- Assets/UnityUIFileBrowser/Scripts/ResetScale.cs
@@ -1,24 +1,0 @@
-﻿// Author: Salvis Poišs (poisins92@gmail.com)
-// Feel free to use and modify (and leave some credits :) )!
-
-using UnityEngine;
-using System.Collections;
-
-using UnityEngine.UI;
-
-/// <summary>
-/// Reset object's scale
-/// </summary>
-[DisallowMultipleComponent]
-public class ResetScale : MonoBehaviour {
-
-	// Use this for initialization
-    public void Start()
-    {
-        if (gameObject.GetComponent<RectTransform>() != null)
-        {
-            gameObject.GetComponent<RectTransform>().localScale = new Vector3(1, 1, 1);
-        }
-    }
-	
-}
