+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,1 @@
-m_EditorVersion: 5.3.5f1
-m_StandardAssetsVersion: 0
+m_EditorVersion: 5.5.0p2
